requirements.txt
numpy>=2.0.0
pandas>=2.0.0
plotly>=5.0.0
pydeck>=0.8.0
streamlit>=1.30.0
Pillow>=10.0.0
